chore(providers): Remove hard-coded interface
