+++ languages/javascript/templates/codeblocks/interface.mjs
@@ -1,0 +1,3 @@
+interface ${name} {
+    ${methods}
+}
