+++ src/shared/modules.mjs
@@ -84,6 +84,83 @@ const getProviderInterfaceMethods = (capability, json) => {
 }
   
 
+function getProviderInterface(capability, module) {
+    module = JSON.parse(JSON.stringify(module))
+    const iface = getProviderInterfaceMethods(capability, module).map(method => localizeDependencies(method, module, null, { mergeAllOfs: true }))
+  
+    iface.forEach(method => {
+      const payload = localizeDependencies(getPayloadFromEvent(method), module)
+      const focusable = method.tags.find(t => t['x-allow-focus'])
+  
+      // remove `onRequest`
+      method.name = method.name.charAt(9).toLowerCase() + method.name.substr(10)
+  
+      method.params = [
+        {
+          "name": "parameters",
+          "schema": payload.properties.parameters
+        },
+        {
+          "name": "session",
+          "schema": {
+            "type": focusable ? "FocusableProviderSession" : "ProviderSession"
+          }
+        }
+      ]
+  
+      let exampleResult = null
+  
+      if (method.tags.find(tag => tag['x-response'])) {
+        const result = method.tags.find(tag => tag['x-response'])['x-response']
+  
+        method.result = {
+          "name": "result",
+          "schema": result
+        }
+  
+        if (result.examples && result.examples[0]) {
+          exampleResult = result.examples[0]
+        }
+      }
+      else {
+        method.result = {
+          "name": "result",
+          "schema": {
+            "const": null
+          }
+        }
+      }
+  
+      method.examples = method.examples.map( example => (
+        {
+          params: [
+            {
+              name: "parameters",
+              value: example.result.value.parameters
+            },
+            {
+              name: "session",
+              value: {
+                correlationId: example.result.value.correlationId
+              }
+            }
+          ],
+          result: {
+            name: "result",
+            value: exampleResult
+          }
+        }
+      ))
+  
+      // remove event tag
+      method.tags = method.tags.filter(tag => tag.name !== 'event')
+    })
+  
+  
+    return iface
+  }
+  
+
 const addMissingTitles = ([k, v]) => {
     if (v && !v.hasOwnProperty('title')) {
         v.title = k
@@ -1074,7 +1151,7 @@ export {
     hasMethodAttributes,
     getMethodAttributes,
     getMethods,
-    getProviderInterfaceMethods,
+    getProviderInterface,
     getProvidedCapabilities,
     getSetterFor,
     getSubscriberFor,
